New packages: linux5.8-pds, linux5.8-pds-zen, and linux5.8-pds-zen2
Inspired by: https://aur.archlinux.org/packages/linux-pds/

Additional patches taken from:

https://aur.archlinux.org/cgit/aur.git/tree/0005-glitched-pds.patch?h=linux-pds
https://aur.archlinux.org/cgit/aur.git/tree/0005-v5.8_undead-pds099o.patch?h=linux-pds

And for the linux5.8-pds-zen/linux5.8-pds-zen2 packages only:

https://github.com/graysky2/kernel_gcc_patch/blob/master/enable_additional_cpu_optimizations_for_gcc_v10.1%2B_kernel_v5.8%2B.patch#L15-L25

I had to rename the two PDS patches so that they applied in the correct order.

The only change I make to the vanilla void kernel template is to not
use sed to change CONFIG_LOCALVERSION; I need it to be what it is. And
I suppose you could count only supporting x86_64 as a change; I never
built this for another arch.